actors: split event serialize benchmark
